co-location spectral initialisers
