+++ packages/insomnia/src/ui/components/codemirror/base-imports.ts
@@ -57,7 +57,6 @@ import 'codemirror-graphql/variables/mode';
 import './modes/nunjucks';
 import './modes/curl';
 import './modes/openapi';
-import './lint/openapi';
 import './lint/javascript-async-lint';
 import './lint/json-lint';
 import './extensions/autocomplete';
